requirements.txt
# 微信公众号文章自动订阅系统依赖
# WeChat Public Account Article Subscriber Dependencies

# 核心依赖
requests>=2.28.0
PyYAML>=6.0

# HTML解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 图片处理（可选，用于WebP转JPG）
Pillow>=10.0.0

# HTTP优化
urllib3>=2.0.0

# 数据库（SQLite为内置，无需安装）

# 进程管理
psutil>=5.9.0

# 测试依赖
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
coverage>=7.3.0

# 开发依赖
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
